requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
SQLAlchemy==2.0.28
watchdog==3.0.0
python-dotenv==1.0.1
ftputil==5.0.4

